+++ cpp/src/parquet/CMakeLists.txt
@@ -347,6 +347,7 @@ add_parquet_test(arrow-test
                  SOURCES
                  arrow/arrow_reader_writer_test.cc
                  arrow/arrow_schema_test.cc
+                 arrow/arrow_statistics_test.cc
                  test_util.cc)
 
 add_parquet_test(arrow-internals-test
