+++ cpp/src/parquet/arrow/arrow_reader_writer_test.cc
@@ -3895,42 +3895,6 @@ TEST_P(TestArrowWriteDictionary, Statistics) {
 INSTANTIATE_TEST_SUITE_P(WriteDictionary, TestArrowWriteDictionary,
                          ::testing::Values(ParquetDataPageVersion::V1,
                                            ParquetDataPageVersion::V2));
-
-TEST(TestArrowWriteNestedDictionary, TestNullCount) {
-  std::shared_ptr<::arrow::Array> dictionary = ArrayFromJSON(::arrow::utf8(), R"(["z"])");
-  std::shared_ptr<::arrow::Array> indices =
-      ArrayFromJSON(::arrow::int32(), R"([0, null, 0])");
-  std::shared_ptr<::arrow::Array> offsets =
-      ArrayFromJSON(::arrow::int32(), R"([null, 0, null, null, null, 3])");
-  ASSERT_OK_AND_ASSIGN(std::shared_ptr<::arrow::Array> dict_encoded,
-                       ::arrow::DictionaryArray::FromArrays(indices, dictionary));
-  // values looks like [null, ["z", null, "z"], null, null, null]
-  ASSERT_OK_AND_ASSIGN(
-      std::shared_ptr<Array> values,
-      ::arrow::ListArray::FromArrays(*offsets, *dict_encoded, default_memory_pool()));
-  std::shared_ptr<::arrow::Schema> schema =
-      ::arrow::schema({::arrow::field("values", values->type())});
-  std::shared_ptr<::arrow::Table> table = ::arrow::Table::Make(schema, {values});
-
-  std::shared_ptr<::arrow::ResizableBuffer> serialized_data = AllocateBuffer();
-  auto out_stream = std::make_shared<::arrow::io::BufferOutputStream>(serialized_data);
-  std::unique_ptr<FileWriter> writer;
-  ASSERT_OK(FileWriter::Open(*table->schema(), default_memory_pool(), out_stream,
-                             default_writer_properties(),
-                             default_arrow_writer_properties(), &writer));
-  ASSERT_OK(writer->WriteTable(*table, std::numeric_limits<int64_t>::max()));
-  ASSERT_OK(writer->Close());
-  if (!out_stream->closed()) {
-    ASSERT_OK(out_stream->Close());
-  }
-
-  auto buffer_reader = std::make_shared<::arrow::io::BufferReader>(serialized_data);
-  auto parquet_reader = ParquetFileReader::Open(std::move(buffer_reader));
-  std::shared_ptr<FileMetaData> metadata = parquet_reader->metadata();
-  std::shared_ptr<Statistics> stats = metadata->RowGroup(0)->ColumnChunk(0)->statistics();
-  EXPECT_EQ(stats->num_values(), 2);
-  EXPECT_EQ(stats->null_count(), 5);
-}
 // ----------------------------------------------------------------------
 // Tests for directly reading DictionaryArray
 
@@ -4225,108 +4189,6 @@ TEST(TestArrowReadDeltaEncoding, DeltaBinaryPacked) {
   GTEST_SKIP() << "Test needs CSV reader";
 }
 #endif
-
-struct StatisticsTestParam {
-  std::shared_ptr<::arrow::Table> table;
-  int expected_null_count;
-  // This is the non-null count and not the num_values in the page headers.
-  int expected_value_count;
-  std::string expected_min;
-  std::string expected_max;
-};
-
-class ParameterizedStatisticsTest : public ::testing::TestWithParam<StatisticsTestParam> {
-};
-
-std::string GetManyEmptyLists() {
-  std::string many_empty_lists = "[";
-  for (int i = 0; i < 2000; ++i) {
-    many_empty_lists += "[],";
-  }
-  many_empty_lists += "[1,2,3,4,5,6,7,8,null]]";
-  return many_empty_lists;
-}
-
-TEST_P(ParameterizedStatisticsTest, NoNullCountWrittenForRepeatedFields) {
-  std::shared_ptr<::arrow::ResizableBuffer> serialized_data = AllocateBuffer();
-  auto out_stream = std::make_shared<::arrow::io::BufferOutputStream>(serialized_data);
-  std::unique_ptr<FileWriter> writer;
-  ASSERT_OK(FileWriter::Open(*GetParam().table->schema(), default_memory_pool(),
-                             out_stream, default_writer_properties(),
-                             default_arrow_writer_properties(), &writer));
-  ASSERT_OK(writer->WriteTable(*GetParam().table, std::numeric_limits<int64_t>::max()));
-  ASSERT_OK(writer->Close());
-  if (!out_stream->closed()) {
-    ASSERT_OK(out_stream->Close());
-  }
-
-  auto buffer_reader = std::make_shared<::arrow::io::BufferReader>(serialized_data);
-  auto parquet_reader = ParquetFileReader::Open(std::move(buffer_reader));
-  std::shared_ptr<FileMetaData> metadata = parquet_reader->metadata();
-  std::shared_ptr<Statistics> stats = metadata->RowGroup(0)->ColumnChunk(0)->statistics();
-  EXPECT_EQ(stats->null_count(), GetParam().expected_null_count);
-  EXPECT_EQ(stats->num_values(), GetParam().expected_value_count);
-  ASSERT_TRUE(stats->HasMinMax());
-  EXPECT_EQ(stats->EncodeMin(), GetParam().expected_min);
-  EXPECT_EQ(stats->EncodeMax(), GetParam().expected_max);
-}
-
-INSTANTIATE_TEST_SUITE_P(
-    StatsTests, ParameterizedStatisticsTest,
-    ::testing::Values(
-        StatisticsTestParam{/*table=*/Table::Make(
-                                ::arrow::schema({::arrow::field("a", ::arrow::utf8())}),
-                                {ArrayFromJSON(::arrow::utf8(), R"(["1", null, "3"])")}),
-                            /*expected=*/1, /* empty list counts as null as well */
-                            /*expected_value_count=*/2,
-                            /*expected_min=*/"1",
-                            /*expected_max=*/"3"},
-
-        StatisticsTestParam{
-            /*table=*/Table::Make(
-                ::arrow::schema({::arrow::field("a", ::arrow::list(::arrow::utf8()))}),
-                {ArrayFromJSON(::arrow::list(::arrow::utf8()),
-                               R"([["1"], [], null, ["1", null, "3"]])")}),
-            /*expected_null_count=*/3, /* empty list counts as null as well */
-            /*expected_value_count=*/3,
-            /*expected_min=*/"1",
-            /*expected_max=*/"3"},
-        StatisticsTestParam{
-            /*table=*/Table::Make(
-                ::arrow::schema({::arrow::field("a", ::arrow::int64())}),
-                {ArrayFromJSON(::arrow::int64(), R"([1, null, 3, null])")}),
-            /*expected_null_count=*/2, /* empty list counts as null as well */
-            /*expected_value_count=*/2,
-            /*expected_min=*/std::string("\x1\0\0\0\0\0\0\0", 8),
-            /*expected_max=*/std::string("\x3\0\0\0\0\0\0\0", 8)},
-        StatisticsTestParam{
-            /*table=*/Table::Make(
-                ::arrow::schema({::arrow::field("a", ::arrow::list(::arrow::utf8()))}),
-                {ArrayFromJSON(::arrow::list(::arrow::utf8()),
-                               R"([["1"], [], ["1", "3"]])")}),
-            /*expected_null_count=*/1, /* empty list counts as null as well */
-            /*expected_value_count=*/3,
-            /*expected_min=*/"1",
-            /*expected_max=*/"3"},
-        StatisticsTestParam{
-            /*table=*/Table::Make(
-                ::arrow::schema({::arrow::field("a", ::arrow::list(::arrow::int64()))}),
-                {ArrayFromJSON(::arrow::list(::arrow::int64()),
-                               R"([[1], [], null, [1, null, 3]])")}),
-            /*expected_null_count=*/3, /* empty list counts as null as well */
-            /*expected_value_count=*/3,
-            /*expected_min=*/std::string("\x1\0\0\0\0\0\0\0", 8),
-            /*expected_max=*/std::string("\x3\0\0\0\0\0\0\0", 8)},
-
-        StatisticsTestParam{
-            /*table=*/Table::Make(
-                ::arrow::schema({::arrow::field("a", ::arrow::list(::arrow::int64()),
-                                                false)}),
-                {ArrayFromJSON(::arrow::list(::arrow::int64()), GetManyEmptyLists())}),
-            /*expected_null_count=*/2001, /* empty list counts as null as well */
-            /*expected_value_count=*/8,
-            /*expected_min=*/std::string("\x1\0\0\0\0\0\0\0", 8),
-            /*expected_max=*/std::string("\x8\0\0\0\0\0\0\0", 8)}));
 
 }  // namespace arrow
 }  // namespace parquet
